FEAT: State is now an array of string values. Enables more than one value to be included.
Create method now accepts an access node parameter which can be used to set a difference access node at the end of the storage operation to the one used to commence the storage operation. This is used in SWAN to use a difference access node (SWAN Operator) for the CMP and the Publisher.

Gzip compression has been added to all responses ahead of performance testing.