FEAT: Modified SWIFT so that the solution can post a message back to the parent window after the storage operation completes as well as change the navigation of the current window. There is a also a feature to turn off the graphical user interface for displaying progress. This has been done to provide more flexibility for callers to operate SWIFT in a pop-up window or an iFrame.
A new flag byte has been added to the Operation data structure to support the bits for, post message on complete indicator and also to turn the user interface on an off. The remaining 6 bits are not yet used.

The new query string parameters to use these features are "postMessageOnComplete" and "displayUserInterface". They take the value "true" or "false".

The BundleTimeout configuration field has been changed to StorageOperationTimeout because Bundle is no longer a word used in the SWIFT vocabulary.