FEAT/OPTIM: Added a flag to enable/disable the homenode check in storage operations and modified to enable empty values to be provided to make the storage operation read only.
The flag useHomeNode=true/false can be used to enable or disable the home node check in the storage operation. As the home node check is verifying that all the cookies held are within the current time limit and are not empty it does not check to determine if the values match the current storage operation. This is okay when a read request is occurring, however when the UIP triggers an update and new data is xplicitly being written to the network the presence of value isn't enough to verify only the home node can be used. This flag enables the caller to bypass the check and is likely to be used for write operations.

The operation now enables empty values to be provided rather than requiring a default value to be provided if a value does not already exist for the key in the network. This enables the SWIFT layer to avoid writing cookies for empty values  but still be informed that the value for the key is required. The following key provided to SWIFT as a parameter woudl read the value for "key" and address any conflicts by choosing the most recent value.

"key>="

No cookie would be written for key as it has no value.

This change ensures that no cookies are written to the web browser until after values have been provided. For SWAN this is important for the User Interface Provider to ensure their dialogue has been displayed and the user enter their preferences before anything is written.