BUG/OPTIM: Fixed a bug that prevent the settings warning appearing when cookies are not available.
The storage operation now records the browser warning percentage as a float 32 value.

If the value is not 0 then the browser warning is displayed.

An empty cookie is written for any storage requests that do not involve cookies being written. This is done for operational reasons to determine if the web browser supports cookies and inform the browser warning if they are not.

Timeouts in seconds are now integers. Functions are provided to return them as durations.