FEAT: Added a new attribue of a storage node called cookie domain. This can be used to provide a different domain the origin domain of the URL to store cookies in. The feature is needed to support the storage of cookies in the parent domain where an operator is assigned a sub domain. i.e. operator-a.exmaple.com and operator-b.example.com store the raw SWAN data in the domain example.com.
Made the scrambler and secrets optional. If the scrambler is not provided the table name appears without modification in the URL and cookie path. If the secret is not provided then the data is compressed but not encrypted. These are needed when a common cookie jar is being used by multiple storage nodes and per node encryption is not desirable.

Moved the compress and decompress methods into a new file called compress.go.

Register method now takes three additional parameters, secret, scramble and cookie domain to support configuration for the new features.